requirements.txt
rich
argparse
python-magic
pandas
xlrd
